src/app/shared/components/error-page/error-page.component.ts
import { NgClass } from '@angular/common';
import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
import { MatButton } from '@angular/material/button';
import { Router, RouterLink } from '@angular/router';

@Component({
  selector: 'app-error-page',
  standalone: true,
  imports: [MatButton, RouterLink, NgClass],
  templateUrl: './error-page.component.html',
  styleUrl: './error-page.component.scss',
  changeDetection: ChangeDetectionStrategy.OnPush
})
export class ErrorPageComponent implements OnInit {
  currentUrl: string = '';
  constructor(private route: Router) {}

  ngOnInit() {
    this.currentUrl = this.route.url;
  }
}
